refactor(order/repository): modularize in-memory repository implementation
